Move all compiler bits to compiler
